requirements.txt
opencv-python==4.8.1.78
scikit-learn==1.3.2
scikit-image==0.22.0
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
numpy==1.24.3
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2